LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.16-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Return statements are only valid inside functions
S7.5.2_A1.16

PASS Expected parsing failure

TEST COMPLETE
